pyproject.toml
[project]
name = "mcp-basics"
version = "0.1.0"
description = "mcp basics"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio>=4.9.0",
    "fastmcp>=2.2.7",
]
